# server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for ChatGameLab - an AI-powered interactive game platform",
        "title": "ChatGameLab API",
        "contact": {
            "name": "ChatGameLab Team",
            "url": "https://chatgamelab.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/system/settings": {
            "get": {
                "description": "Returns the global system settings",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Get system settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/obj.SystemSettings"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates the global system settings (admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Update system settings",
                "parameters": [
                    {
                        "description": "Settings to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.UpdateSystemSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/obj.SystemSettings"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/system/settings/free-use-key": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sets or clears the global free-use API key (admin only).\nThe admin's own API key will be used directly.\nPass null apiKeyId to clear.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Set system free-use API key",
                "parameters": [
                    {
                        "description": "API Key ID to set (null to clear)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.SetFreeUseApiKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/obj.SystemSettings"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httpx.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Machine-readable error code",
                    "type": "string"
                },
                "message": {
                    "description": "Human-readable error message",
                    "type": "string"
                },
                "type": {
                    "description": "Deprecated: use Code instead",
                    "type": "string"
                }
            }
        },
...
        "obj.SystemSettings": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "defaultAiQualityTier": {
                    "description": "ultimate server-wide fallback (e.g. \"medium\")",
                    "type": "string"
                },
                "freeUseAiQualityTier": {
                    "description": "tier for system free-use key, nil = use default",
                    "type": "string"
                },
                "freeUseApiKeyId": {
                    "type": "string"
                },
                "freeUseApiKeyName": {
                    "type": "string"
                },
                "freeUseApiKeyPlatform": {
                    "type": "string"
                },
                "freeUseApiKeyWorking": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "modifiedAt": {
                    "type": "string"
                },
                "promptConstraintU13": {
                    "description": "constraint for users aged 13-17",
                    "type": "string"
                },
                "promptConstraintU18": {
                    "description": "constraint for users aged 18+",
                    "type": "string"
                }
            }
        },
...
        "routes.SetFreeUseApiKeyRequest": {
            "type": "object",
            "properties": {
                "apiKeyId": {
                    "type": "string"
                }
            }
        },
...
        "routes.UpdateSystemSettingsRequest": {
            "type": "object",
            "properties": {
                "defaultAiQualityTier": {
                    "type": "string"
                },
                "freeUseAiQualityTier": {
                    "type": "string"
                },
                "promptConstraintU13": {
                    "type": "string"
                },
                "promptConstraintU18": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token in format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}